refactor: profile exports for actions and stores in desktop
