+++ packages/desktop/components/ProposalsDetails.svelte
@@ -1,13 +1,4 @@
 <script lang="ts">
-    import { onMount } from 'svelte'
-
-    import { Button, KeyValueBox, Text, ProposalsDetailsButton } from '@ui'
-    import { ButtonSize, FontWeight } from '@ui/enums'
-
-    import { selectedAccount } from '@core/account'
-    import { localize } from '@core/i18n'
-    import { activeProfileId } from '@core/profile'
-
     import { IProposalsDetails } from '@contexts/governance/interfaces'
     import {
         participationOverviewForSelectedAccount,
@@ -20,8 +11,13 @@
         getNumberOfVotedProposals,
         getNumberOfVotingProposals,
     } from '@contexts/governance/utils'
-
-    import { openPopup, PopupId } from '@desktop/auxiliary/popup'
+    import { selectedAccount } from '@core/account'
+    import { localize } from '@core/i18n'
+    import { activeProfileId } from '@core/profile/stores'
+    import { PopupId, openPopup } from '@desktop/auxiliary/popup'
+    import { Button, KeyValueBox, ProposalsDetailsButton, Text } from '@ui'
+    import { ButtonSize, FontWeight } from '@ui/enums'
+    import { onMount } from 'svelte'
 
     let details = <IProposalsDetails>{
         totalProposals: null,
