+++ packages/desktop/components/modals/AccountSwitcherModal.svelte
@@ -4,7 +4,8 @@
     import { selectedAccount } from '@core/account/stores'
     import { formatCurrency, localize } from '@core/i18n'
     import { getMarketAmountFromAssetValue } from '@core/market/utils'
-    import { activeProfile, getBaseToken, visibleActiveAccounts } from '@core/profile'
+    import { activeProfile, visibleActiveAccounts } from '@core/profile/stores'
+    import { getBaseToken } from '@core/profile/actions'
     import { formatTokenAmountBestMatch, selectedAccountAssets } from '@core/wallet'
     import { PopupId, openPopup } from '@desktop/auxiliary/popup'
     import { AccountSwitcherMenuItem, FontWeight, Modal, Text, TextType } from '@ui'
