+++ packages/desktop/components/popups/AddProposalPopup.svelte
@@ -1,12 +1,13 @@
 <script lang="ts">
     import { showNotification } from '@auxiliary/notification/actions'
-    import { closePopup, openPopup, PopupId } from '@desktop/auxiliary/popup'
-    import { registerProposalsForAccounts, registeredProposalsForSelectedAccount } from '@contexts/governance'
+    import { registeredProposalsForSelectedAccount, registerProposalsForAccounts } from '@contexts/governance'
     import { selectedAccount } from '@core/account'
     import { handleError } from '@core/error/handlers/handleError'
     import { localize } from '@core/i18n'
-    import { activeAccounts, updateActiveAccountPersistedData } from '@core/profile'
+    import { updateActiveAccountPersistedData } from '@core/profile/actions'
+    import { activeAccounts } from '@core/profile/stores'
     import { truncateString } from '@core/utils/string'
+    import { closePopup, openPopup, PopupId } from '@desktop/auxiliary/popup'
     import type { Auth } from '@iota/wallet'
     import { Button, Checkbox, NodeInput, Text, TextInput, TextType } from '@ui'
     import { HTMLButtonType } from '@ui/enums'
