+++ packages/desktop/components/popups/AliasConfirmationPopup.svelte
@@ -1,19 +1,19 @@
 <script lang="ts">
-    import { Button, KeyValueBox, Text, FontWeight, TextType } from '@ui'
     import { selectedAccount, updateSelectedAccount } from '@core/account'
+    import { processAndAddToActivities } from '@core/activity/utils'
+    import { handleError } from '@core/error/handlers/handleError'
     import { localize } from '@core/i18n'
-    import { checkActiveProfileAuth, getBaseToken } from '@core/profile'
+    import { checkActiveProfileAuth, getBaseToken } from '@core/profile/actions'
     import {
-        convertBech32ToHexAddress,
-        formatTokenAmountPrecise,
         EMPTY_HEX_ID,
         UNLOCK_CONDITION_GOVERNOR_ADDRESS,
         UNLOCK_CONDITION_STATE_CONTROLLER_ADDRESS,
+        convertBech32ToHexAddress,
+        formatTokenAmountPrecise,
     } from '@core/wallet'
     import { closePopup } from '@desktop/auxiliary/popup'
+    import { Button, FontWeight, KeyValueBox, Text, TextType } from '@ui'
     import { onMount } from 'svelte'
-    import { handleError } from '@core/error/handlers/handleError'
-    import { processAndAddToActivities } from '@core/activity/utils'
 
     export let _onMount: (..._: any[]) => Promise<void> = async () => {}
