+++ packages/desktop/components/popups/ManageVotingPowerPopup.svelte
@@ -1,16 +1,16 @@
 <script lang="ts">
-    import { Button, Text, TextHint, AssetAmountInput } from '@ui'
-    import { HTMLButtonType, TextType } from '@ui/enums'
+    import { setVotingPower } from '@contexts/governance/actions'
+    import { isAccountVoting } from '@contexts/governance/utils'
     import { selectedAccount } from '@core/account/stores'
     import { handleError } from '@core/error/handlers'
-    import { setVotingPower } from '@contexts/governance/actions'
     import { localize } from '@core/i18n'
     import { checkActiveProfileAuth } from '@core/profile/actions'
+    import { activeProfile } from '@core/profile/stores'
     import { convertToRawAmount, visibleSelectedAccountAssets } from '@core/wallet'
-    import { closePopup, openPopup, PopupId, popupState } from '@desktop/auxiliary/popup'
+    import { PopupId, closePopup, openPopup, popupState } from '@desktop/auxiliary/popup'
+    import { AssetAmountInput, Button, Text, TextHint } from '@ui'
+    import { HTMLButtonType, TextType } from '@ui/enums'
     import { onMount } from 'svelte'
-    import { isAccountVoting } from '@contexts/governance/utils'
-    import { activeProfile } from '@core/profile'
 
     export let _onMount: (..._: any[]) => Promise<void> = async () => {}
     export let newVotingPower: string = undefined
