+++ packages/desktop/components/popups/MintNftFormPopup.svelte
@@ -1,20 +1,20 @@
 <script lang="ts">
-    import { onMount } from 'svelte'
-    import { Button, Error, FontWeight, OptionalInput, Text, TextInput, TextType, TooltipIcon } from '@ui'
-    import { closePopup, openPopup, PopupId } from '@desktop/auxiliary/popup'
     import { BaseError } from '@core/error/classes'
     import { handleError } from '@core/error/handlers/handleError'
     import { localize } from '@core/i18n'
+    import { composeUrlFromNftUri } from '@core/nfts'
     import { MimeType } from '@core/nfts/types'
-    import { isValidUri } from '@core/utils/validation'
+    import { fetchWithTimeout } from '@core/nfts/utils/fetchWithTimeout'
+    import { getNetworkHrp } from '@core/profile/actions'
+    import { HttpHeader } from '@core/utils'
     import { validateBech32Address } from '@core/utils/crypto'
+    import { isValidUri } from '@core/utils/validation'
+    import { IMintNftDetails } from '@core/wallet'
     import { TokenStandard } from '@core/wallet/enums'
     import { mintNftDetails, setMintNftDetails } from '@core/wallet/stores'
-    import { IMintNftDetails } from '@core/wallet'
-    import { fetchWithTimeout } from '@core/nfts/utils/fetchWithTimeout'
-    import { composeUrlFromNftUri } from '@core/nfts'
-    import { HttpHeader } from '@core/utils'
-    import { getNetworkHrp } from '@core/profile'
+    import { PopupId, closePopup, openPopup } from '@desktop/auxiliary/popup'
+    import { Button, Error, FontWeight, OptionalInput, Text, TextInput, TextType, TooltipIcon } from '@ui'
+    import { onMount } from 'svelte'
 
     export let _onMount: (..._: any[]) => Promise<void> = async () => {}
