+++ packages/desktop/views/dashboard/send-flow/views/components/StorageDepositButton.svelte
@@ -1,10 +1,10 @@
 <script lang="ts">
-    import { Text, Icon, Modal, Toggle, FontWeight } from '@ui'
     import { Icon as IconEnum } from '@auxiliary/icon'
-    import { fade } from 'svelte/transition'
-    import { formatTokenAmountPrecise } from '@core/wallet'
-    import { getBaseToken } from '@core/profile'
     import { localize } from '@core/i18n'
+    import { getBaseToken } from '@core/profile/actions'
+    import { formatTokenAmountPrecise } from '@core/wallet'
+    import { FontWeight, Icon, Modal, Text, Toggle } from '@ui'
+    import { fade } from 'svelte/transition'
 
     export let storageDeposit: number
     export let giftStorageDeposit: boolean
