+++ packages/desktop/views/dashboard/settings/views/network/LocalProofOfWork.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
     import { localize } from '@core/i18n'
     import { toggleLocalPowInClientOptions } from '@core/network'
-    import { activeProfile } from '@core/profile'
+    import { activeProfile } from '@core/profile/stores'
     import { NetworkSettingsRoute } from '@core/router'
     import { Checkbox } from '@ui'
     import SettingsSection from '../SettingsSection.svelte'
