+++ packages/desktop/views/dashboard/settings/views/profile/ChangeProfileName.svelte
@@ -1,8 +1,9 @@
 <script lang="ts">
-    import { Button, Input, Text, HTMLButtonType, ButtonSize, TextType } from '@ui'
-    import { localize } from '@core/i18n'
     import { showNotification } from '@auxiliary/notification'
-    import { activeProfile, updateActiveProfile, validateProfileName } from '@core/profile'
+    import { localize } from '@core/i18n'
+    import { validateProfileName } from '@core/profile'
+    import { activeProfile, updateActiveProfile } from '@core/profile/stores'
+    import { Button, ButtonSize, HTMLButtonType, Input, Text, TextType } from '@ui'
 
     let newName = $activeProfile?.name
     let error = ''
