+++ packages/desktop/views/onboarding/views/network-setup/views/ChooseNetworkView.svelte
@@ -5,7 +5,7 @@
     import { IS_MOBILE } from '@core/app'
     import { localize } from '@core/i18n'
     import { NetworkId, getDefaultClientOptions, getDefaultPersistedNetwork } from '@core/network'
-    import { profiles } from '@core/profile'
+    import { profiles } from '@core/profile/stores'
     import features from '@features/features'
     import { Animation, OnboardingButton, Text, TextType } from '@ui'
     import { onMount } from 'svelte'
