+++ packages/desktop/views/update-stronghold/views/ChangePasswordView.svelte
@@ -4,11 +4,12 @@
     import { onboardingProfile, updateOnboardingProfile } from '@contexts/onboarding'
     import { handleError } from '@core/error/handlers'
     import { localize } from '@core/i18n'
-    import { MAX_STRONGHOLD_PASSWORD_LENGTH, unlockStronghold } from '@core/profile'
+    import { MAX_STRONGHOLD_PASSWORD_LENGTH } from '@core/profile'
     import { initialiseProfileManager } from '@core/profile-manager/actions'
     import { changeStrongholdPassword } from '@core/profile-manager/api'
     import { profileManager } from '@core/profile-manager/stores'
     import { buildProfileManagerOptionsFromProfileData } from '@core/profile-manager/utils'
+    import { unlockStronghold } from '@core/profile/actions'
     import { activeProfile, updateActiveProfile } from '@core/profile/stores'
     import { PASSWORD_REASON_MAP } from '@core/stronghold'
     import { Animation, Button, PasswordInput, Text, TextHint } from '@ui'
