+++ packages/shared/src/lib/contexts/governance/actions/initializeParticipationOverviews.ts
@@ -1,4 +1,4 @@
-import { activeAccounts } from '@core/profile'
+import { activeAccounts } from '@core/profile/stores'
 import { ParticipationOverview } from '@iota/wallet'
 import { get } from 'svelte/store'
 import { DEFAULT_PARTICIPATION_OVERVIEW } from '../constants'
