+++ packages/shared/src/lib/core/profile/index.ts
@@ -1,5 +1,3 @@
-export * from './stores'
-export * from './actions'
 export * from './constants'
 export * from './errors'
 export * from './interfaces'
